pyproject.toml
[tool.poetry]
name = "qa guru 19 hw"
version = "0.1.0"
description = "Qa guru 19 hw"
authors = ["Maksim Kudaev <kudaev.m@nakukop.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^7.2.1"
selene = "2.0.0b16"
dynaconf = "^3.1.11"
requests = "^2.28.2"
pytest-voluptuous = "^1.2.0"
voluptuous = "^0.13.1"
allure-pytest = "^2.12.0"
curlify = "^2.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
log_cli = 1
log_cli_level = "INFO"
log_cli_format = "%(asctime)s %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
addopts = " -vv -s --alluredir=allure-result --allure-no-capture "
